requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Cython==3.0.11
Deprecated==1.2.14
et-xmlfile==1.1.0
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.141.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.65.4
grpcio-status==1.62.3
httplib2==0.22.0
idna==3.7
lxml==5.3.0
mysql-connector-python==9.0.0
numpy==2.0.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pdf2image==1.17.0
pefile==2023.2.7
pikepdf==9.1.1
pillow==10.4.0
poppler-utils==0.1.0
proto-plus==1.24.0
protobuf==4.25.4
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32-ctypes==0.2.3
requests==2.32.3
rsa==4.9
setuptools==72.2.0
six==1.16.0
SQLAlchemy==2.0.32
tk==0.1.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
wheel==0.44.0
wrapt==1.16.0

